fix: store xxh3 hash as TEXT because its too big to be stored as int

+++ db/postgres/postgres_test.go
@@ -562,7 +562,7 @@ func TestCheckGraded(t *testing.T) {
 
 	hasher := xxh3.New()
 	hasher.WriteString("joe" + courses[0].Code + professors[0].UUID)
-	hash := int64(hasher.Sum64())
+	hash := hasher.Sum64()
 
 	graded, err := TestDB.checkGraded(hash)
 	if err != nil {
